pyproject.toml
[tool.poetry]
name = "disnake-bot-template-postgres"
version = "0.1"
description = "A template for speeding up development of discord bots using disnake and postgresql."
authors = ["Exenifix"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.27.0"
exencolorlogs = "^1.1.1"
disnake = "^2.7.0"
exenenv = "^1.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
ruff = "^0.0.199"
isort = "^5.11.4"
black = "^22.12.0"
python-dotenv = "^0.21.0"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
